requirements.txt
# Core data science
numpy>=1.24.0,<2.0.0
pandas>=2.0.0
scipy>=1.10.0

# Time-series forecasting
statsmodels>=0.14.0
prophet>=1.1.4

# Deep learning
torch>=2.0.0
torchmetrics>=1.0.0

# Machine learning utilities
scikit-learn>=1.3.0

# API
fastapi>=0.100.0
uvicorn[standard]>=0.23.0
pydantic>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Visualization (for notebooks/evaluation)
matplotlib>=3.7.0
seaborn>=0.12.0

# HTTP client (for LLM API calls)
httpx>=0.24.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0

# Linting
ruff>=0.1.0
